Upgrade PokemonCard to container
