+++ stephanie-zeng/src/containers/SearchBar.js
